requirements.txt
flask
requests
gmqtt
uvloop
paho-mqtt
